Refactor. Production server code
