--- server/.gitignore
@@ -1,11 +1,0 @@
-# The node modules forlder
-node_modules
-configurations
-
-yarn.lock
-
-.env
-
-personalTokens.json
-
-uploads
