--- server/README.md
@@ -1,5 +1,0 @@
-# WCE Connects: Backend
-#### *Establishing long-standing bond between the alumni, current students and teahers.*
-
-Directory Structure:
-![Screenshot](./images/Directory-Structure.jpg?raw=true "The Directory Structure")
