--- server/models/Conversation.js
@@ -1,12 +1,0 @@
-const mongoose = require("mongoose");
-
-const ConversationSchema = new mongoose.Schema(
-  {
-    members: {
-      type: Array,
-    },
-  },
-  { timestamps: true }
-);
-
-module.exports = mongoose.model("Conversation", ConversationSchema);
