--- server/models/Event.js
@@ -1,32 +1,0 @@
-const mongoose = require('mongoose');
-
-const eventSchema = new mongoose.Schema({
-    title: {
-        type: String,
-        required: true
-    },
-    description: {
-        type: String,
-        required: true
-    },
-    date: {
-        type: Date,
-        required: true
-    },
-    location: {
-        type: String,
-        required: true
-    },
-    image: {
-        type: String,
-        required: true
-    },
-    creator: {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'User',
-        required: true
-    }
-}, {timestamps: true});
-
-
-module.exports = mongoose.model('Event', eventSchema);
