--- server/models/Message.js
@@ -1,18 +1,0 @@
-const mongoose = require("mongoose");
-
-const MessageSchema = new mongoose.Schema(
-  {
-    conversationId: {
-      type: String,
-    },
-    sender: {
-      type: String,
-    },
-    text: {
-      type: String,
-    },
-  },
-  { timestamps: true }
-);
-
-module.exports = mongoose.model("Message", MessageSchema);
