+++ server/models/user.js
@@ -1,12 +1,4 @@
-const Joi       = require('joi'),
-    mongoose    = require('mongoose'),
-    jwt         = require('jsonwebtoken');
-
-// Configure dotenv dependancy to access credentials and private keys 
-require('dotenv').config();
-
-// I don't know what it does
-Joi.objectId = require('joi-objectid')(Joi);
+const mongoose    = require('mongoose');
 
 // Schema for the short profile of user
 const userSchema = new mongoose.Schema({
@@ -133,39 +125,7 @@ const userSchema = new mongoose.Schema({
     }
 });
 
-// Method to get jwt token (From objectID and jwt private key)
-// userSchema.methods.genAuthToken = function () {
-//     return jwt.sign({ _id: this._id }, process.env.WC_jwtPrivateKey);
-// };
-
 // Creating mongoDB model from defined schema
 const User = mongoose.model('User', userSchema);
 
-// Joi validation
-// function validateUser(newUser){
-//     const schema = Joi.object({
-//         credentials : {
-//             email : Joi.string().min(10).max(255).required().email(),
-//             password : Joi.string().min(4).max(1024).required(),
-//         },
-//         info : {
-//             first_name : Joi.string().min(3).max(100).required(),
-//             last_name : Joi.string().min(3).max(100).required(),
-//             user_role : Joi.string().required()
-//         } 
-//     });
-//     return schema.validate(newUser);
-// }
-
-// Joi validation
-// function validateParams(params)
-// {
-//     const schema = Joi.object({
-//         id : Joi.objectId()
-//     });
-//     return schema.validate(params);
-// }
-
 module.exports.User = User;
-// module.exports.validateUser = validateUser;
-// module.exports.validateParams = validateParams;
