+++ server/package.json
@@ -1,29 +1,17 @@
 {
-  "name": "sem2project-backend",
-  "version": "0.0.0",
-  "private": true,
+  "name": "restapi.wce.ac.in",
+  "version": "1.0.0",
+  "description": "Mininulistic backend for activity registration form",
+  "main": "app.js",
   "scripts": {
-    "start": "nodemon app.js"
+    "start": "node app.js"
   },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
     "bcrypt": "^5.0.1",
-    "config": "^3.3.6",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "joi": "^17.4.0",
-    "joi-objectid": "^3.0.1",
-    "jsonwebtoken": "^8.5.1",
-    "lodash": "^4.17.21",
-    "mongoose": "^5.12.8",
-    "multer": "^1.4.3",
-    "nodemailer": "^6.6.0",
-    "socket.io": "^4.1.2",
-    "xls-to-json-lc": "^0.3.4",
-    "xlsx-to-json-lc": "^0.5.0"
-  },
-  "devDependencies": {
-    "concurrently": "^6.2.0",
-    "nodemon": "^2.0.7"
+    "mongoose": "^6.0.13",
+    "cors": "*"
   }
 }
