--- server/routes/events.js
@@ -1,13 +1,0 @@
-
-
-
-
-
-// Create Event: Only allowed for admin
-// POST: /events
-
-// Update Event: Only allowed for admin
-
-// Delete Event: Only allowed for admin
-
-// Read Event: 1. Pagination 2. By ID
