--- server/socket/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "socket",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "nodemon": "^2.0.7",
-    "socket.io": "^4.1.2"
-  }
-}
